+++ packages/platform-sdk-egld/src/networks/egld/mainnet.ts
@@ -4,7 +4,7 @@ const network: Coins.CoinNetwork = {
 	id: "egld.mainnet",
 	type: "live",
 	name: "Mainnet",
-	explorer: "https://explorer.elrond.com/",
+	explorer: "https://explorer.elrond.com",
 	currency: {
 		ticker: "EGLD",
 		symbol: "EGLD",
@@ -17,7 +17,10 @@ const network: Coins.CoinNetwork = {
 	},
 	featureFlags: {
 		Client: {
+			transaction: true,
+			transactions: true,
 			wallet: true,
+			broadcast: true,
 		},
 		Identity: {
 			address: {
@@ -32,6 +35,11 @@ const network: Coins.CoinNetwork = {
 			keyPair: {
 				mnemonic: true,
 			},
+		},
+		Link: {
+			block: true,
+			transaction: true,
+			wallet: true,
 		},
 	},
 	transactionTypes: ["transfer"],