+++ packages/platform-sdk-egld/src/services/client.test.ts
@@ -1,0 +1,50 @@
+import "jest-extended";
+
+import nock from "nock";
+
+import { createConfig } from "../../test/helpers";
+import { TransactionData, WalletData } from "../dto";
+import { ClientService } from "./client";
+
+let subject: ClientService;
+
+beforeEach(async () => (subject = await ClientService.__construct(createConfig())));
+
+afterEach(() => nock.cleanAll());
+
+beforeAll(() => nock.disableNetConnect());
+
+describe("ClientService", function () {
+	test("#transaction", async () => {
+		nock(/.+/)
+			.get("/v1.0/transaction/c2e6e2c75357b7d69d735d5ce7d7e9a77291477d0a11ba158b5cf39317398f66")
+			.reply(200, require(`${__dirname}/../../test/fixtures/client/transaction.json`));
+
+		const result = await subject.transaction("c2e6e2c75357b7d69d735d5ce7d7e9a77291477d0a11ba158b5cf39317398f66");
+
+		expect(result).toBeInstanceOf(TransactionData);
+	});
+
+	test("#transactions", async () => {
+		nock(/.+/)
+			.get("/v1.0/address/erd17uy64y9zw8zd4xz5d2deqmxfxkk3zfuj0jh24k0s5jqhet3pz0esng60j7/transactions")
+			.reply(200, require(`${__dirname}/../../test/fixtures/client/transactions.json`));
+
+		const result = await subject.transactions({
+			addresses: ["erd17uy64y9zw8zd4xz5d2deqmxfxkk3zfuj0jh24k0s5jqhet3pz0esng60j7"],
+		});
+
+		expect(result).toBeObject();
+		expect(result.items()[0]).toBeInstanceOf(TransactionData);
+	});
+
+	test("#wallet", async () => {
+		nock(/.+/)
+			.get("/v1.0/address/erd17uy64y9zw8zd4xz5d2deqmxfxkk3zfuj0jh24k0s5jqhet3pz0esng60j7")
+			.reply(200, require(`${__dirname}/../../test/fixtures/client/wallet.json`));
+
+		const result = await subject.wallet("erd17uy64y9zw8zd4xz5d2deqmxfxkk3zfuj0jh24k0s5jqhet3pz0esng60j7");
+
+		expect(result).toBeInstanceOf(WalletData);
+	});
+});