+++ CHANGELOG.md
@@ -5,6 +5,57 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 3.0.12 - 2020-02-24
+
+### Added
+
+-   **[ETH]** Retrieve fees from `ethgas.watch` (c259a012, @faustbrian)
+
+### Changed
+
+-   Store plugins with an internal UUID (5fd0f0ec, @faustbrian)
+
+## 3.0.10 - 2020-02-24
+
+### Fixed
+
+-   **[EGLD]** Link `blocks` to `miniblocks` (d0b02ae2, @faustbrian)
+-   **[EGLD]** Treat transaction timestamp as unix (2009aeb9, @faustbrian)
+
+## 3.0.8 - 2020-02-24
+
+### Fixed
+
+-   **[EGLD]** Normalise fees to satoshi (a2ccaf62, @faustbrian)
+
+## 3.0.7 - 2020-02-24
+
+### Fixed
+
+-   **[EGLD]** Normalise numeric values to satoshis (4d4e3def, @faustbrian)
+
+## 3.0.6 - 2020-02-24
+
+### Fixed
+
+-   **[EGLD]** Expect `egld.mainnet` or `egld.testnet` as networks (706c9076, @faustbrian)
+-   **[EGLD]** Expect address or list of addresses for `ClientService#transactions` (56734116, @faustbrian)
+-   Adjust validation of plugins (fbd078fb, @dated)
+
+## 3.0.3 - 2020-02-24
+
+### Added
+
+-   **[EGLD]** Implement `ClientService#wallet` (6d7a0b15, @faustbrian)
+
+### Changed
+
+-   **[EGLD]** Adjust DTOs and some API paths (756ff46c, @faustbrian)
+
+### Fixed
+
+-   Use `EGLD` as export name (003c67e1, @faustbrian)
+
 ## 3.0.0 - 2020-02-23
 
 See https://github.com/jeanlescure/string-crypto#v2-breaking-changes for breaking change details.